Rename typing.py to type_util.py
typing.py clashes with the standard library typing module
