+++ tests/test_tracing.py
@@ -19,7 +19,7 @@ from monkeytype.tracing import (
     get_func,
     trace_calls,
 )
-from monkeytype.typing import NoneType
+from monkeytype.type_util import NoneType
 
 # avoid a hard dependency on Cython
 try:
